move remaining fs ops
