+++ cli/js/ops/timers.ts
@@ -8,3 +8,12 @@ export function stopGlobalTimer(): void {
 export async function startGlobalTimer(timeout: number): Promise<void> {
   await sendAsync("op_global_timer", { timeout });
 }
+
+interface NowResponse {
+  seconds: number;
+  subsecNanos: number;
+}
+
+export function now(): NowResponse {
+  return sendSync("op_now");
+}
